a  demo of product interface
This is a new request I uploaded. The product interface I made based on the original framework could not run, so at the request of my tutor, I completed the creation of a small product interface demonstration application using typescript and react. Please note that the application is just a demonstration program, which contains a video of the program's operation process.